move remaining canister code
