+++ benchmark-canisters/src/vec.rs
@@ -1,7 +1,8 @@
-use crate::{benchmark, BenchResult, Random};
+use crate::Random;
 use ic_cdk_macros::query;
 use ic_stable_structures::storable::Blob;
 use ic_stable_structures::{DefaultMemoryImpl, StableVec, Storable};
+use profiler::{benchmark, BenchResult};
 use tiny_rng::{Rand, Rng};
 
 #[query]
